move pointPen to fontTools.pens subpackage; keep ufoLib.pointPen alias
